refactor: Refactor several something in all layers

+++ api/httpserver/middleware/check_is_valid_user_id.go
@@ -12,12 +12,11 @@ import (
 )
 
 func CheckIsValidUserID(next echo.HandlerFunc) echo.HandlerFunc {
-
 	return func(c echo.Context) error {
 		id := c.Param("id")
 		idFromToken := claim.GetClaimsFromEchoContext[authservice.Claims](c, configs.AuthMiddlewareContextKey).UserID
 
-		idFromTokenConvertToSTR := strconv.FormatUint(uint64(idFromToken), 10)
+		idFromTokenConvertToSTR := strconv.FormatUint(idFromToken, 10)
 
 		if id != idFromTokenConvertToSTR {
 			return echo.NewHTTPError(http.StatusBadRequest, echo.Map{
